Enables RAG support for chat and refactors embedder utils
Enables retrieval-augmented generation (RAG) integration in chat endpoints, allowing document context to be used in chat responses when attachments are present. Refactors embedding model utility functions into a dedicated module for better code organization. Enhances user experience by providing more relevant, context-aware answers.
