+++ backend/rag/default_embedder.py
@@ -1,12 +1,10 @@
 import os
 import PyPDF2
 import hashlib
-from pathlib import Path
 from typing import List
 from langchain_text_splitters import RecursiveCharacterTextSplitter
 from langchain_community.vectorstores import FAISS
 from langchain_huggingface import HuggingFaceEmbeddings
-from huggingface_hub import snapshot_download
 
 
 class DefaultRAG:
@@ -99,38 +97,3 @@ class DefaultRAG:
 
         documents = self.vector_store.similarity_search(query, k=k)
         return [doc.page_content for doc in documents]
-
-
-# Check if the embedding model is downloaded
-def check_embedding_model(
-    model_name="sentence-transformers/all-MiniLM-L6-v2",
-):
-    home = Path.home()
-    cache_dir = str(home / ".cache" / "huggingface" / "hub")
-    try:
-        snapshot_download(
-            repo_id=model_name,
-            local_dir=None,  # Use default cache location
-            local_dir_use_symlinks=False,
-            revision="main",
-            repo_type="model",
-            cache_dir=cache_dir,
-            local_files_only=True,  # Only check local files, don't download
-        )
-        return True
-    except Exception:
-        return False
-
-
-# Download the embedding model
-def download_embedding_model(model_name="sentence-transformers/all-MiniLM-L6-v2"):
-    if check_embedding_model(model_name):
-        return {"status": "success"}
-    try:
-        HuggingFaceEmbeddings(
-            model_name=model_name,
-            model_kwargs={"device": "cpu"},
-        )
-        return {"status": "success"}
-    except Exception as e:
-        return {"status": "error", "message": str(e)}
